Refactor repository code into specialized modules
Decompose the large `extended.go` file into smaller, logical files
(`queries_auth.go`, `queries_feed.go`, `queries_locations.go`,
`queries_media.go`, `queries_migrations.go`, `queries_posts.go`,
`queries_system.go`, `queries_tags.go`, `queries_timeline.go`) and
update corresponding test files. Move storage statistics query logic to
the repository layer.
